Remove __start__ node from stream output
