+++ langgraph/pregel/debug.py
@@ -1,23 +1,20 @@
-from collections import deque
 from pprint import pformat
 from typing import Any, Iterator, Mapping
 
-from langchain_core.runnables import Runnable
 from langchain_core.utils.input import get_bolded_text, get_colored_text
 
 from langgraph.channels.base import BaseChannel, EmptyChannelError
+from langgraph.pregel.types import PregelExecutableTask
 
 
-def print_step_start(
-    step: int, next_tasks: list[tuple[Runnable, Any, str, deque[tuple[str, Any]]]]
-) -> None:
+def print_step_start(step: int, next_tasks: list[PregelExecutableTask]) -> None:
     n_tasks = len(next_tasks)
     print(
         f"{get_colored_text('[langgraph/step]', color='blue')} "
         + get_bolded_text(
             f"Starting step {step} with {n_tasks} task{'s' if n_tasks > 1 else ''}. Next tasks:\n"
         )
-        + "\n".join(f"- {name}({pformat(val)})" for _, val, name, _ in next_tasks)
+        + "\n".join(f"- {name}({pformat(val)})" for name, val, _, _, _ in next_tasks)
     )
